Update patch set 1
Patch Set 1: Code-Review+2

The changes look good to me.
But, it is generally recommended to push smaller related fixes individually rather than grouping unrelated ones in a single push.
That makes a logical cherry-picking or reverting possible without the need for manual intervention. 
But, I will go ahead and approve this because of the harmless nature of the unrelated fixes.

Patch-set: 1
Label: Code-Review=+2