Update patch set 1
Patch Set 1: Code-Review-1

It is a good idea to insert the file that needs update without depending on the infinispan upstream.

But, I worry about the tighter (and not enforceable) dependency on Infinispan 5.3.0 release. What if someone in the future upgrades the Infinispan to later versions ?
Since there is no enforcement, this will sneak in and override any runtime fixes on this file for the later version.

So, though the intention is good, it can cause some human error in the future. Hence requesting to add some sort of build time enforcement atleast with a clear comment pointing to this Gerrit and the File (org/infinispan/interceptors/distribution/BaseDistributionInterceptor.java) that calls for this enforcement.

Patch-set: 1
Label: Code-Review=-1