Silent false positive, and enhance real error message
- With gerrit https://git.opendaylight.org/gerrit/#/c/1366/ i have
introduced a log error, to spot inconsistent hashcode but indeed it
turn out is getting exercised also during a normal code path,
basically when:
 <command orginator> --> <key owner> --> <all the other members>

all the other members when getting the replication from the non-key
owner will end up in this case. So that cannot be actually considered
an error.

Change-Id: Id38b45066164cb4998fdcdbb356bdf7af720b8c5
Signed-off-by: Giovanni Meo <gmeo@cisco.com>