Update patch set 1
Patch Set 1: Code-Review-1

The binary result from isLocal() is inadequate and does give a false impression on the locality of the node.
Infact that is the reason we deprecated it and replaced it with getLocalityStatus().

We have already seen the issues with isLocal() on various scenarios. Overloading it and removing deprecation notice on this method will cause more confusion and harm.

Patch-set: 1
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1