Newcomers that register with TopologyManager should get all the existing edges
- When an application wants to get topology updates by topology
manager, it may come after the topology manager has started to push
them around, in that case the app should retrieve the exisiting
topology. Some applications may not do that so topology manager can
still take care of this lazy one by pushing a snapshot of the existing
edges as soon as the listener is registered.

Change-Id: I919350b2ddfddfaa9186e09bd5167d7d810e3bcd
Signed-off-by: Giovanni Meo <gmeo@cisco.com>