Add HTTP Client library and refactor Northbound IntegrationTest
Create a new commons.httpclient library for bundles that need
to function as an HTTP client.  Refactor the existing NB
IntegrationTest code to use this new library.

Ammended - library is now a profile of apache httpclient classes
to ensure things like timeouts are set on default and resources
are correctly released when finished.

Change-Id: I76f513b77700213752f4183e917fe239a9c324ff
Signed-off-by: Ryan Moats <rmoats@us.ibm.com>