Simple Sanity Test to validate that all the bundles in a distribution are active and resolved
The idea behind the Sanity Test is to ultimately fail a build when the test fails. The test would
launch run.sh but with a slightly modified config.ini where the sanitytest bundle would be injected
as a framework bundle.

The test waits for 30 seconds for the container to finish loading all the bundles. It then gets all the
bundles from the BundleContext and checks if they are ACTIVE or RESOLVED. If they are neither it prints
out the name of the bundle and it's state and returns an error code.

Change-Id: Ia6905318f1cf63e852791ba107193094a83fddf1
Signed-off-by: Moiz Raja <moraja@cisco.com>