Update patch set 2
Patch Set 2:

Before merging that commit, make sure that this is announced on controller-dev.

The problem is - any bundle in controller project is automatically inserted into distribution, so developer's needs to update distribution if their bundle is not to be part of OSGi Runtime.

There are several artefacts which are not meant to go into distribution e.g.:
  - archetypes
  - checkstyle
  - config/yang-jmx-generator-plugin

Patch-set: 2
Reviewer: Gerrit User 21 <21@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0