Update patch set 2
Patch Set 2:

The current distribution pom mixes 'build the modules' with 'assemble the distribution'.  We'd like to eventually separate it.  Until then, we should add <excludes /> to exclude things that should be built but should not be included.

This patch though doesn't actually *use* the sanitytest yet, it just provides the bundle.  So no harm in taking this patch.

Patch-set: 2