Allow Topology DB to keep track of production switches
ISSUE: Discovery service is able to discover any non SDN enable
switch directly conencted to the controlled SDN switches which
advertise itself through LLDP. It is a nice feature to be able
to see on the topology to which non-SDN switches the controlled
network nodes are conencted to.
Change in https://git.opendaylight.org/gerrit/#/c/1450/ breaks this feature.

CHANGE: Limit the validation check introduced in the above commit
to the edge head end point only.

Change-Id: I081f4320a73eea9bf2a39ca80717771d9c7adfb3
Signed-off-by: Alessandro Boch <aboch@cisco.com>