Update patch set 1
Patch Set 1: Code-Review-2

Tony as it's this will break the distribution, if you want to move the dependencies out of opendaylight/commons/opendaylight/pom.xml make sure to have a mechanism in opendaylight/distribution/opendaylight/pom.xml to collect the dependencies and put in the plugins directory. Your diffs don't have that now.

Patch-set: 1
Label: Code-Review=-2