Update patch set 9
Patch Set 9:

Overall, this looks good. In general, I think keeping module configuration as close as is possible to the actual module is good. This does that for dependencies. I'm  not sure that it will make things easy to have to put each module into the root pom.xml and distribution pom.xml, but as long as this doesn't cross git repos, I think that's probably fine.

From a more nit-oriented point of view: The junit dependency sometimes has a version and other times it doesn't. Similarly it sometimes has a scope and other times doesn't. Is there a way to standardize that? Could we have one copy of each dependency in a dependencies.xml somewhere and then reference them by name so that the blocks aren't everywhere or is that hard?

Patch-set: 9