Update patch set 1
Patch Set 1: Code-Review+2

Normally I would not self review this.  But as currently the config system does not load into the controller, and it is blocking the inclusion of a sanitytest case that would prevent that in the future, I am doing so in this exceptional case.

Patch-set: 1
Reviewer: Gerrit User 9 <9@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+2
Label: SUBM=+1