Update patch set 2
Patch Set 2:

Due to the importance of this change and the potential impact this might have on other projects, was it considered to have this ADSAL -> MDSAL migration done as part of a topic branch ?

Also, the Integration tests should be beafed up to cover all the APIs.

Hence IMHO, a topic branch would be a better fit to make sure the changes works and then find its way to the master.

Patch-set: 2
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0