Update patch set 5
Patch Set 5:

Hi Ryan,

I understand that this is building on top of an existing sanity test script. Hence commenting on this code-review is not appropriate (as we discussed in the IRC extensively).

If this is not blocking your progress in windows, I would suggest you to make progress without this patch.

Prasanth and few others have other cleaner approach/ideas with regards to the Sanity tests. Lets wait on that rather than piling up on a fragile test script.

-Madhu

Patch-set: 5