Create change
Patch-set: 1
Change-id: Iad74e8f4ea84ab6115908fa39859227afdfdf53f
Subject: Modified the setting of JAVA_HOME for linux to be more generic. The new code sets the value with the assumption that the user has java somewhere in their path and that it is set to the JAVA version they want to use. JAVA_HOME is then set based on that location.
Branch: refs/heads/master
Commit: 126a368e726c0e8abc9427645b98291270a12df9