Update patch set 5
Patch Set 5: Code-Review+1

Based on the various discussions, it is fair to get this commit merged in so that the windows side of things are handled properly.
The identified issues on the Sanity tests can be addressed independent of this fix.

Patch-set: 5
Reviewer: Gerrit User 11 <11@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1