Update patch set 1
Patch Set 1:

Sure. I missed the /osgi/ on the URL.
That brings in an important point about this URL construction.

We need a consistent NB-URL for any resources exposed by the Controller. So, I would rather keep the URL as /controller/nb/v2/jolokia to be consistent with the rest of the controller REST API.

Can we share the same HTTP Bridge and still achieve this ?
If not, i would prefer to have a separate bundle to just expose the same service from a different URL like the one mentioned above.

Please let me know.

Patch-set: 1