Update patch set 1
Patch Set 1:

the /osgi/ in the URL is where my problem is :(
jolokia being a jmx REST bridge, i would prefer it to be
/controller/nb/v2/ as in any other REST API base.

Also, I agree that the httpservice-bridge should not have /nb/ in it due to the obvious reasons.

So, given these facts, I think it is prefered to have a separate bundle like that one in this patch and maybe I will experiment with listForHttpService to make sure that Jolokia registers on its own and not to the httpservice-bridge.

Patch-set: 1