Update patch set 2
Patch Set 2:

I have an overall design comment with this push.

In my opinion, this is a well defined functionality that can stand on its own bundle making use of the FRM and Statistics Manager APIs.

It is more like an Application of these bundles and can independently evolve without being tied with the FRM or Statistics manager bundles.

Can you please consider this design comment ?
If not, can you please give the design intent behind overloading FRM and Statistics manager with this functionality ?

Patch-set: 2