Update patch set 2
Patch Set 2: Code-Review+2

It's true that with the newly added IFMR API getFlowsForNode() the evaluation if a inconsistency is there can be done outside of FRM context.
But it is also true that this is a building block for a future FRM internal consistency checker.

Regarding the changes to Statistics Manager, they are needed regardless of the CC evaluation is done outside or inside FRM, as there is no way today to retrieve the non cached flow statistics for a node controlled by another controller in the cluster.

Patch-set: 2