Update patch set 5
Patch Set 5:

Hi Muthu,

Sorry about the naming comment again. I still think getControllers() is more intuitive and friendly than the proposed getCareFor getCareForClusterMembers.
We are in the controller project and IConnectionManager is infact used by NSFs and applications. This has no direct relationships with Southbound plugins such as BGPCEP or OF.

IMHO calling it a controller is the right thing to do
(After all there are many existing controllers that are not OF or specific protocol based).

-Madhu

Patch-set: 5
Label: Verified=0