Add env vars to set the controller working directory.
The run.sh makes an assumption that the current working directory is
writeable and owned by the user running the command. These assumptions
don't hold when running the controller as a service such as when it
was installed via an RPM package.

Added two env variables to specify the run directory and the output
directory:

ODL_BASEDIR: working directory
ODL_DATADIR: output directory

By default the run.sh will continue to use the current working directory
and will only use the env vars if they are set.

Change-Id: I85dda38353f632e84409d91ac4dfc4a44c2d457e
Signed-off-by: Sam Hague <shague@redhat.com>