Update patch set 1
Patch Set 1:

Alissa... do you know the trick for making a Draft globally visible?  The reason I didn't use the Draft feature is that its invisible except to the folks you explicitly as as reviewers, and I want global visibility for this so we can all talk about it.  If you can point me to how to do a globabally visible draft I'd be most grateful as I'd find that to be hugely useful :)

Patch-set: 1