Update patch set 6
Patch Set 6: Code-Review+1

I definitively like that this is shifting the test to be an IT test and can be tuned via knobs. The fact it doesn't run on windows i also reflected and yes we probably need to find a way to support run.bat -start and -stop (-jmx is trivial).

Patch-set: 6
Reviewer: Gerrit User 19 <19@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1