Update patch set 1
Patch Set 1:

Hi Giovanni,

I see your point.

Ganymed is quite matured Java SSH library but is written in object / procedural manner using standard socket. which means there is no way to extend any class of library. Or at least I did not found a way. If there is a need for SSH to be part of Netty pipeline, easiest think to do is to replace ganymed created socket with other one. This needs to be done transparently and with as few patches to ganymed sourcecode as possible. So instead of extending class, ganymed is patched at two places to accept precreated (NETTY) socket instead of creating own one. This is reason why we need patched sources in ODL.

That precreated socket is virtual socket (part of commit as well) created out of underlying netty handlers. So from the Ganymed point of view ganymed is working with socket interface with streams, but from Netty point of view there are just handlers. So we create handler which acts as virtual socket, input/output stream handlers which at Ganymed acts as regular I/O streams but at Netty they are regular handlers.

This way:
1. ganymed has transparent small change
2. virtually any OIO library is able to run in NIO environment.

Hope explanation make sense.

Patch-set: 1
Label: Verified=0