Update patch set 1
Patch Set 1:

The NetConfClient constructor throws InterruptedException.
It is very unexpected when creating a class to experience exception - especially this type, that can cause the object to be not created. And it will be very hard to troubleshoot it.
It is better to refrain in the constructor itself from things like creating threads, working with sockets, etc.

Patch-set: 1
Reviewer: Gerrit User 739 <739@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0