Update patch set 1
Patch Set 1:

Milos, I am wondering - if this is a general threadpool service to be used by the controller, why it has to be implemented via yang and not "regular" Java? What is the benefit? 
I understand it is a data modeling language for netconf protocol, hence it is not clear to me why a basic and core serivce such as threadpool, that is not related to any network protocol by itself has to be implemented in a data modelling language intended for a network protocol.
Keep in mind, most of Java developers are not familiar with yang, hence less chance to review and maintain in properly.
Will be glad to get your explanation on this. Thank you.

Patch-set: 1
Reviewer: Gerrit User 739 <739@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0