Initial commit for the ClusteredDataStore
The ClusteredDataStore provides a mechanism for the MD-SAL Broker to share data across a cluster. It can be used by other components/services as well to share data.

The ClusteredDataStore will use the ClusteringServices interface to create a cache and read/write data to it.

This only includes the structure of the project.

Change-Id: I37b24c8a91e6dc69815566bc0acd6dde561cbc29
Signed-off-by: Moiz Raja <moraja@cisco.com>