Update patch set 3
Patch Set 3:

Robert definitively much better, but i'd like to propose that you further split the fix in these 3 parts:

1) The introduction of the ganymede ssh wrapper and the two classes you plan on changing but before actually doing the changes
2) A patch that just show the two classes changed (this is very easy and stays for the record on what you actually changed in ganymede ssh
3) A patch with your code that uses ganymede-ssh

You already have the different pieces, just this way it shows in the logs the whole evolution and stays for the record also afterward.
Also on #1 i suggest you to add an enforcer rule to make sure this compiles only with the version of ganymed-ssh you are starting from. Given you are caching source code and that comes from that revision of the ganymed-ssh it's a safety measure to avoid someone later on upgrade the library and runs in weird issues.
Hopes it make sense.

Thanks,
Giovanni

Patch-set: 3
Label: Verified=0