Update patch set 5
Patch Set 5:

it is better to do such changes as renames and moving of the files in a separate patch. it is not directly related to the new functionality introduced in this patch. and it will be a smaller and more concrete patch, easier review, less changes to apply at once ...

Patch-set: 5