Update patch set 1
Patch Set 1:

(4 comments)

I'm not super-familiar with the YANG modeling yet, but most of this is self-explanatory. My only other concerns is that we probably want to be able to have more than one node-connector, DL addr, NW addr, and maybe even VLAN ID for each host. We may also want to encode which of those are bound to each other.

For instance, a host might have more than one NIC and have a unique IP and MAC for each or they might have the same IP and MAC on both. It might also be the case that they have two different MACs, but each on different VLANs.

In the process of doing all of that, we also come to the part where we need to figure out if we want to encode a unique identifier for each host in the model?

Patch-set: 1
Reviewer: Gerrit User 48 <48@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0