Reduce logged messages during build and concentrate only on the ones needed for debugging build errors
- Create a maven artifact that contains a default settings for
logback.xml so it can be reused as dependency to make sure some
plugins don't generate too much verbose output
- Use the artifact
org.opendaylight.controller:commons.logback_settings as UT dependency
so to make sure only in case of errors the follow UT spit messages:
 	+ netconf-impl
	+ config-netconf-connector
- Remove duplicated logback.xml where not needed
- Remove logback.xml and space_and_tabs_checks.xml where no longer needed
- Added logback.xml as a file on the integration tests using pax-exam,
there a local file seems to be required (unfortunately). Make sure the
pax-exam uses it for the tests where not using it.
- Made sure the sanitytest in distribution only runs when
integrationstests profile is selected
- Fixed an issue in importing in eclipse zeromq module
- Excluded all tests from mdsal unless the integrationtests profile is
selected

Signed-off-by: Giovanni Meo <gmeo@cisco.com>
Change-Id: I9c12965bb3de5c8cba1039992cadbe27026bca2a