Fix discrepancy between command line maven build and the one initiated by a launcher
- m2 launchers are all set with the "Resolve Workspace Artifacts"
which essentially force all the artifacts to be built in the view and
if some is missing they are not downloaded from nexus. This is not
inline with the command line build which is what we use in the
jenkins, so restoring a consistent behavior between command line and
m2 launchers.
- Remove the sample.toaster-it from the distribution pom it's not
needed there.
- Added a new full launcher that will trigger the build also on
integrations tests and docs profiles

Change-Id: I09cf57ffb7771f9d1787b5ff9d24b2b0af25abb6
Signed-off-by: Giovanni Meo <gmeo@cisco.com>