Update patch set 5
Patch Set 5:

(1 comment)

Please do not merge this patch currently.

It contains a hack to turn off sal-compatibility for the purpose of allowing the openflowplugin folks to cherry-pick it in and get unblocked.  When Prasanna pushes a fix that deals with the build breakage, we can remove the pom.xml change and only *then* merge.

Patch-set: 5
Reviewer: Gerrit User 9 <9@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0