Modify the java-client to use the build-helper-maven-plugin
- Current northbound.java-client uses custom install-file and
deploy-file to actually deploy multiple artifacts. Now that has the
issue of telling to which repo the artifact should go, while it's way
simpler to let maven figure out via the distributionManagement. To
rescue in this situation the build-helper-maven-plugin can attach
additional files to an artifact distinguing them via the classifier
selector. This change is to use this feature

Change-Id: Id425ca2d9e744de7cceb0028afd84dfb48ba93a7
Signed-off-by: Giovanni Meo <gmeo@cisco.com>