Topology Manager changes for cross checking Production switches
LLDP messages are received by OpenFlow protocol plugin. This plugin only knows
of OpenFlow switches connected to the Controller, not the other types of
switches. If discovery service in OF plugin can't correlate the sender of
LLDP message to a known switch, it assumes LLDP is sent by a "Production"
type of switch which is not connected to Controller. However, LLDP packet
may have been sent by a non OpenFlow switch which is not known to OF Plugin
but is known to SwitchManager. The LLDP message results in an edgeupdate
message to TopologyManager which then can consult with SwitchManager to see
if the switch previously categorized as Production Switch exists in Switch
Manager database and re-classify it to a known type of switch.

Change-Id: I42b649d099346ac7187037ae2c5f7efd27260293
Signed-off-by: Maurice Qureshi <maquresh@cisco.com>