Update patch set 1
Patch Set 1:

Ed, I have asked Ryan to perform a simple diff to ensure his files are safe, and the only difference is \r.
I didn't see his response on that till now.
I can abandon the patch, thought I don't see any real justification to do so.
Given the amount of work I invested in the patch, and the amount of work that will be required to redo it due to code changes that will enter till the release, I hope someone else will volunteer to redo it after the release.

Patch-set: 1