Update patch set 3
Patch Set 3: Code-Review+2

Alessandro, thanks for taking care of the isIPv6 comment in a more generic way. I still believe we can do better, and express a more generic predicate, but lets keep functional and this code to me looks functional. If we find a better way to express a more generic predicate i think we can then remove the isIP6 for now i agree this is handy and should be used.

Patch-set: 3
Label: Code-Review=+2