Update patch set 1
Patch Set 1:

HashMap initialized this way may need cleanup depending on the how long the lifecycle of the object span. For now i will just merge this as long as i don't have an understanding of how it's used but please make sure the logic holds in case of object being reused.

Patch-set: 1