Fixes for MD-FRM
Patch 1:  Fix MD-FRM to listen to the correct portion of the data tree.
Currently we listen to /flows/flow/ for which we would need to specify a
key to indicate which flow in the tree we wanted to listen to.  Changing to
listen to /flows/ and thus hear commits involving all flows.

Change-Id: Iee020938d201f3df0629b2df13a85ad023c808ac
Signed-off-by: Ed Warnicke <eaw@cisco.com>