DO NOT MERGE YET - Fixes for MD-FRM
Patch 1:  Fix MD-FRM to listen to the correct portion of the data tree.
Currently we listen to /flows/flow/ for which we would need to specify a
key to indicate which flow in the tree we wanted to listen to.  Changing to
listen to /flows/ and thus hear commits involving all flows.

Patch 2:  Unpack Entry... we need the value of the entry which is a DataObject.
Otherwise, the cast fails.

Patch 3: NodeFlow is only used as an input to rpcs towards the
openflowplugin, it does not get filed in the data tree.  Flow gets
filed in the data tree.  Addtionally, commits can come in with multiple
objects, so if we get a commit with other things, we either need to ignore
them or handle them... but blindly casting them will not work.

Patch 4: Change to cope with moving from NodeFlow to Flow.  Fortunately,
NodeFlow is just Flow plus a NodeRef plus a TableRef, so it pretty much just
worked (at least at the typing level... not to running yet).

Patch 5: Fixed an NPE by providing a Collections.<RpcError>emptySet() for
RpcErrors instead of null.

Patch 6: Fixed a failure due to attempt to check with the old AD-SAL
infra for the Node.  This will not work correctly.  Added TODO for
validation of the node.  Also cleaned up substantially handling of
error conditions while I was in the neighborhood.

Patch 7: Something in the current use of ClusteringServices for
validation is breaking things.  At first it looked like
we were just trying to allocate caches before we initialized
out clustering services variable, and thus failing... but at
the end of the day there is more wrong here than make sense
to debug at this time... so commented out.

Change-Id: Iee020938d201f3df0629b2df13a85ad023c808ac
Signed-off-by: Ed Warnicke <eaw@cisco.com>