DO NOT MERGE YET - Fixes for MD-FRM
Patch 1:  Fix MD-FRM to listen to the correct portion of the data tree.
Currently we listen to /flows/flow/ for which we would need to specify a
key to indicate which flow in the tree we wanted to listen to.  Changing to
listen to /flows/ and thus hear commits involving all flows.

Patch 2:  Unpack Entry... we need the value of the entry which is a DataObject.
Otherwise, the cast fails.

Patch 3: NodeFlow is only used as an input to rpcs towards the
openflowplugin, it does not get filed in the data tree.  Flow gets
filed in the data tree.  Addtionally, commits can come in with multiple
objects, so if we get a commit with other things, we either need to ignore
them or handle them... but blindly casting them will not work.

Change-Id: Iee020938d201f3df0629b2df13a85ad023c808ac
Signed-off-by: Ed Warnicke <eaw@cisco.com>