Update patch set 7
Patch Set 7:

(2 comments)

This is a relatively big patch with many changes.
However from the commit message it's not very clear what is the reason for this patch.
It will be great to have more details in commit message for the patch purpose and motivation.

Patch-set: 7
Reviewer: Gerrit User 739 <739@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0