Update patch set 4
Patch Set 4:

Robert, could you explain/add in commit message what is the jsr305 used for and why is it needed as a bundle? 
Also, is it safe to change completely the package? It's not the same packaging vendor, nor the same version, is there anywhere a comparison link to ensure that this is a safe replace?
Thank you.

Patch-set: 4
Reviewer: Gerrit User 739 <739@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0