Update patch set 1
Patch Set 1:

Alissa,

i did inform the original submitter offline. Didn't think to include him as code reviewer, but i will try to add now for the record.
The issue could be detected in the integration testbed because it involves the usage of OpenVswitch, in fact that is where it has been caught. Having a test in the IT or UT wold pose too many requirement on the developer system which may not be satisfied unless you are developing on a Linux system.

Thanks,
Giovanni

Patch-set: 1