Update patch set 9
Patch Set 9:

(1 comment)

The changes done JAX RS looks good to me for the models and plugins code.

But in my view FRM Flow code needs to do following tasks for now:
1. Validate incoming data.
2. Commit to MD-SAL. Call corresponding RPC.
3. Rollback on error.

So for the set of tasks defined above a FLOW FRM file as before should suffice and will be easy to understand for other who are reading the code.

So many classes in FLOW to do 3 tasks looks to me confusing when I see it for understanding purpose.

gerrit 3512 is reference implementation for group without old clustering service.

SO I agree validation should be separated and re-designed into smaller pieces of code.
I agree for different package for flow / group and meter.
I agree for yang model and plugin changes.

Flow FRM is what I feel is confusing and may be required in future when we have to do more things in 2nd phase of commit.
In my view lets hold on to FRM Flow changes for now.

Patch-set: 9
Reviewer: Gerrit User 397 <397@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0