Persist service references as separate MBeans.
Create dummy MBean - ServiceReferenceMXBean - that contains link to implementation of the service as an attribute.
Add support for resolving of SR to dependency resolver. Each time user saves SR, create MBean and
add the interface QName together with user defined reference name to ObjectName of the newly created SR MBean.

Change-Id: I71ddb762e00d9be8b426fb979ff5ade046d46050
Signed-off-by: Tomas Olvecky <tolvecky@cisco.com>