Generate ADDED events for Node and NodeConnector
Components like switch manager need to see an ADDED event come in for Node and NodeConnector before
they accept changes for those elements. This commit first checks whether a Node/NodeConnector are
already known to the DataBrokerService. If they are not known then it notifies the listeners of an ADDED
event otherwise it notifies it of a CHANGED event

Also implemented node removal. NOTE: NodeConnector removed events do not seem to be coming through.

Change-Id: Idcbe7fdce37c16158614f62b56767bed624bc803
Signed-off-by: Moiz Raja <moraja@cisco.com>