Provide an MDSalNodeFactory to construct an MD-SAL node for NB interfaces like the Forwarding Rules Manager
When you try to get flows using the FR NB API for a given node you have to provide both a node type and node id. Since MD_SAL nodes
have a type MD_SAL which is not one of the standard types you need to provide and register a NodeFactory which can construct a Node
for an MD_SAL Node type.

Change-Id: Idc6631f58e248a1d958227312de07faa1fbc3d6e
Signed-off-by: Moiz Raja <moraja@cisco.com>