Upgrade jackson library to version 2.3.0
This change purpose is to use a newer version of the library,
as well as align it with the jackson version used by ovsdb project repo.

Main change areas:
1.Upgrade the jackson from codehaus.jackson version 1.9.x to
fasterxml.jackson version 2.3.0.
change all usages in pom files
and fix the imports in Java classes.
2.Remove the usage of jersey-json library.

Main resource for upgrade information:
http://wiki.fasterxml.com/JacksonUpgradeFrom19To20

Change-Id: I2caa9f94a81af1e5ad0ead90c60c912396b8a390
Signed-off-by: Alissa Bonas <abonas@redhat.com>