Update patch set 9
Patch Set 9:

Hi Alissa,

Thanks for taking care of this change quickly.

Can you please upgrade it to 2.2.2 ?
That is the current version being used in the OVSDB project
and I think that is the latest version used across all other
projects as well.

Other than that, am good. Will merge it once it is changed to 2.2.2 and the NorthboundIT passes.

-Madhu

Patch-set: 9
Label: Code-Review=0