Update patch set 1
Patch Set 1:

Hi Alessandro,

Thanks for your reply.

Here is more information on the nature of the problem:
https://lists.opendaylight.org/pipermail/controller-dev/2013-December/002381.html
This issue was discussed in the mailing list two weeks ago.

I think the change you mentioned does the right thing and solves part of the problem of cross southbound plugins topology discovering. However, there is some remaining part of the problem and this patch tries to fix that.

Thanks,
Paiwei

Patch-set: 1