unitests: Remove extending TestCase from unitest classes.
Extending TestCase which is a Junit3 format, and using Junit4
annotations at the same time in same class causes confusion,
and can also lead to runtime problems.
An example for a problem - putting @Ignore on a test case which starts
with word "test" and is located in a class that extends TestCase will
actually NOT ignore the test because it will be interpreted as Junit3
and not as Junit4.
Removing extending TestCase will make the unitests to be interpreted
as Junit4 style and leave no room for confusion.

Change-Id: I2365f0921d64667a5489c42eeb53adadc45efdaa
Signed-off-by: Alissa Bonas <abonas@redhat.com>