Optimization - calculate the subnet prefix only once.
Before this change, subnet prefix was calculated on every
'isSubnetOf' call. Since the prefix is not changing unless
the subnet address itself changed, it's sufficient to do it only once
and keep it as part of the Subnet object.

Marked it as transient so it will not be serialized - it can be
calculated from the rest of the info of Subnet anytime after
deserialization.

Change-Id: Ie16bf80713820b566864bea0ff46298c93cc2d47
Signed-off-by: Alissa Bonas <abonas@redhat.com>