Update patch set 2
Patch Set 2:

Hi
It seems there is a problem with the commit message, it contains the link to the patch itself which is redundant, and at the same time doesn't provide any information about what is the content of this patch.
Please fix it.
Also, it seems that quite a lot of code is added, however the patch has no tests.

Patch-set: 2
Reviewer: Gerrit User 739 <739@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0