Fixed Node Properties in Adapter - with exceptions noted below
Still need to fix Supported Actions -

Supported Actions present a bit of an issue.  In OF 1.0,
they come in with the GetFeatures message on switch connection.
In OF 1.3, they are per table, and come in with the response to
an empty bodied TableFeatures Multipart request.  So far,
no switch has been observed in the wild supporting the
empty bodied TableFeature message. OVS 2.0 doesn't, CPQD doesn't.

The net-net is that supported Actions are effectively unreportable
for OF 1.3 switches.

Change-Id: I6f4beab167fa1502cacc767eaf95099ba996a343
Signed-off-by: Ed Warnicke <eaw@cisco.com>