Switch from using a Long as FlowId to a Uri
Couple of things you may want to pay attention to,

a. FlowProgrammerAdapter maintains a mapping between a Flow and it's id in a map in infinispan
b. Look at internalAddFlowAsync. If it finds that a flow is in the flow to id mapping cache. It will remove the flow and add it back in. This is what makes the switch add back a flow when it a switch comes back alive.
c. The sync vs async behavior should now be properly done

Change-Id: Id156ebbc7b6e643e4a5960aa735d0812bf5cbe48
Signed-off-by: Moiz Raja <moraja@cisco.com>